backend/pyproject.toml
[build-system]
requires        = ["setuptools>=68", "wheel"]
build-backend   = "setuptools.build_meta"

[project]
name            = "ProjeCT"
description     = "A full-stack application for computed tomography (CT) research."
version         = "0.5.0"
readme          = "README.md"
requires-python = ">=3.10"
dependencies    = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "opengate==10.0.1",
    "pydantic-settings>=2.8.1",
    "pyvista>=0.44.2",
    "sqlalchemy>=2.0.40"
]
